langgraph-platform/pyproject.toml
[project]
name = "14-langserve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "fastmcp>=2.13.0.2",
    "langchain-community>=0.3.8",
    "langchain-core>=0.3.14",
    "langchain-mcp>=0.2.1",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.2.8",
    "langchain-tavily>=0.2.12",
    "langchain-text-splitters>=0.3.2",
    "langgraph==0.6.4",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-sdk>=0.1.38",
    "langsmith>=0.2.7",
    "pymupdf>=1.24.10",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.8.0",
    "tiktoken>=0.7.0",
]
